workflows: Same configs between two IPsec workflows
This commit simply ensures the two IPsec workflows use the same configs.
The configs from the end-to-end test workflow, which seem more
extensive, are used.

Note that this means IPsec+KPR is now covered in up/downgrade tests. It
was only covered in end-to-end tests before. This is now okay because
IPsec+KPR should work in v1.16.

Signed-off-by: Paul Chaignon <paul.chaignon@gmail.com>