WIP: Add support for FIPS compliance
Also restricts all TLS config to FIPS-approved settings.

Note : We (Author of this commit, Datadog Inc. or Cilium Maintainers) are
not making any statements or representations about the suitability of this
code in relation to the FIPS 140 standard. Interested users will have
to evaluate for themselves whether the code is useful for their own
purposes.

Fixes: #23919

Signed-off-by: Hemanth Malla <hemanth.malla@datadoghq.com>